backend/requirements.txt
fastapi
uvicorn
anthropic
numpy
python-dotenv
